+++ rust/crates/iqrah-core/src/exercises/ayah_chain_tests.rs
@@ -165,7 +165,7 @@ impl ContentRepository for MockContentRepo {
         Ok(Vec::new())
     }
 
-    async fn get_word(&self, _word_id: i32) -> anyhow::Result<Option<crate::Word>> {
+    async fn get_word(&self, _word_id: i64) -> anyhow::Result<Option<crate::Word>> {
         Ok(None)
     }
 
@@ -208,7 +208,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_word_translation(
         &self,
-        _word_id: i32,
+        _word_id: i64,
         _translator_id: i32,
     ) -> anyhow::Result<Option<String>> {
         Ok(None)
@@ -292,7 +292,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_morphology_for_word(
         &self,
-        _word_id: i32,
+        _word_id: i64,
     ) -> anyhow::Result<Vec<crate::MorphologySegment>> {
         Ok(Vec::new())
     }
@@ -345,8 +345,8 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_words_batch(
         &self,
-        word_ids: &[i32],
-    ) -> anyhow::Result<std::collections::HashMap<i32, crate::Word>> {
+        word_ids: &[i64],
+    ) -> anyhow::Result<std::collections::HashMap<i64, crate::Word>> {
         let mut result = std::collections::HashMap::new();
         for &id in word_ids {
             if let Some(word) = self.get_word(id).await? {
