+++ rust/crates/iqrah-core/src/exercises/memorization_tests.rs
@@ -287,7 +287,7 @@ mod tests {
             Ok(self.words.get(verse_key).cloned().unwrap_or_default())
         }
 
-        async fn get_word(&self, word_id: i32) -> anyhow::Result<Option<Word>> {
+        async fn get_word(&self, word_id: i64) -> anyhow::Result<Option<Word>> {
             for words in self.words.values() {
                 if let Some(word) = words.iter().find(|w| w.id == word_id) {
                     return Ok(Some(word.clone()));
@@ -335,7 +335,7 @@ mod tests {
 
         async fn get_word_translation(
             &self,
-            _word_id: i32,
+            _word_id: i64,
             _translator_id: i32,
         ) -> anyhow::Result<Option<String>> {
             Ok(None)
@@ -419,7 +419,7 @@ mod tests {
 
         async fn get_morphology_for_word(
             &self,
-            _word_id: i32,
+            _word_id: i64,
         ) -> anyhow::Result<Vec<crate::MorphologySegment>> {
             Ok(vec![])
         }
@@ -472,8 +472,8 @@ mod tests {
 
         async fn get_words_batch(
             &self,
-            word_ids: &[i32],
-        ) -> anyhow::Result<std::collections::HashMap<i32, crate::Word>> {
+            word_ids: &[i64],
+        ) -> anyhow::Result<std::collections::HashMap<i64, crate::Word>> {
             let mut result = std::collections::HashMap::new();
             for &id in word_ids {
                 if let Some(word) = self.get_word(id).await? {
