+++ rust/crates/iqrah-core/src/exercises/translate_phrase_tests.rs
@@ -14,7 +14,7 @@ use std::collections::HashMap;
 struct MockContentRepo {
     texts: HashMap<i64, String>,                        // node_id -> text
     verse_translations: HashMap<(String, i32), String>, // (verse_key, translator_id) -> translation
-    word_translations: HashMap<(i32, i32), String>,     // (word_id, translator_id) -> translation
+    word_translations: HashMap<(i64, i32), String>,     // (word_id, translator_id) -> translation
     words: HashMap<String, Vec<Word>>,                  // verse_key -> words
 }
 
@@ -160,7 +160,7 @@ impl ContentRepository for MockContentRepo {
         Ok(self.words.get(verse_key).cloned().unwrap_or_default())
     }
 
-    async fn get_word(&self, _word_id: i32) -> anyhow::Result<Option<Word>> {
+    async fn get_word(&self, _word_id: i64) -> anyhow::Result<Option<Word>> {
         Ok(None)
     }
 
@@ -206,7 +206,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_word_translation(
         &self,
-        word_id: i32,
+        word_id: i64,
         translator_id: i32,
     ) -> anyhow::Result<Option<String>> {
         Ok(self
@@ -293,7 +293,7 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_morphology_for_word(
         &self,
-        _word_id: i32,
+        _word_id: i64,
     ) -> anyhow::Result<Vec<MorphologySegment>> {
         Ok(vec![])
     }
@@ -346,8 +346,8 @@ impl ContentRepository for MockContentRepo {
 
     async fn get_words_batch(
         &self,
-        word_ids: &[i32],
-    ) -> anyhow::Result<std::collections::HashMap<i32, crate::Word>> {
+        word_ids: &[i64],
+    ) -> anyhow::Result<std::collections::HashMap<i64, crate::Word>> {
         let mut result = std::collections::HashMap::new();
         for &id in word_ids {
             if let Some(word) = self.get_word(id).await? {
